Add job_summary to job data
Updated scrapers to add job_summary data. Jobs with PDF descriptions get
instructions to see PDF document. A Community of Friends job summaries
are not easily scrapable so the url to the job page is the summary.
Union Station Homeless Services job summaries are scraped and are set
correctly. All other orgs are not done yet so they are set to the job
page urls.